requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
# psycopg2
python-dotenv
twilio
uuid
Pillow